qt: make table row count conversions explicit (#287)
Make the two masternode table row-count conversions to Qt's int API explicit and alphabetize the Qt test includes. This is a three-line cleanup on top of the merged hidden-refresh change; it does not change table behavior.

Validation: git diff --check passed. PR CI passed: build, 680 unit cases with no errors, MasternodeListTests 8/8, leak-check cases 4/4, and the functional gate (ALL Passed). Qt totals include initialization/cleanup. No separate local rebuild was performed.

+++ src/qt/masternodelist.cpp
@@ -391,8 +391,8 @@ bool MasternodeList::updateDIP3List()
 
     // Allocate once, avoiding repeated shifts of all preceding rows. Fill in
     // reverse traversal order, as insertRow(0) did, to preserve sort ties.
-    ui->tableWidgetMasternodesDIP3->setRowCount(rows.size());
-    int row_index = rows.size();
+    ui->tableWidgetMasternodesDIP3->setRowCount(static_cast<int>(rows.size()));
+    int row_index = static_cast<int>(rows.size());
     for (auto& row : rows) {
         --row_index;
         for (int column = 0; column <= COLUMN_PROTX_HASH; ++column) {
